LLVM cmplog factoring custom Instruction iterator with added restriction.
